% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{corrplot}
\alias{corrplot}
\title{Plot modelled correlations between species}
\usage{
corrplot(
  object,
  species = NULL,
  plotfun = "intervals",
  nrow = NULL,
  widths = NULL,
  ...
)
}
\arguments{
\item{object}{The jsdmStanFit model object}

\item{species}{Which species should be included - this plots the correlations
between the specified species and all other species, so by default if all
species are included this will result in a plot per species with all other
species there and thus duplicate entries across all the plots}

\item{plotfun}{Which plotting function from bayesplot to use, by default
\code{"intervals"}}

\item{nrow}{How many rows the grid of plots has}

\item{widths}{Whether the widths of the different rows of plots should vary}

\item{...}{Other arguments passed to mcmc_plot}
}
\value{
An object of class \code{"bayesplot_grid"}, for more information see
\code{\link[bayesplot:bayesplot_grid]{bayesplot::bayesplot_grid()}}
}
\description{
Plot modelled correlations between species
}
