% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_funs.R
\name{cov_matern}
\alias{cov_matern}
\alias{cov_sq_exp}
\alias{cov_exp}
\title{Calculate covariance function based on a given data matrix}
\usage{
cov_matern(x, sq_eta, rho, delta = 1e-05, nu05)

cov_sq_exp(x, sq_eta, rho, delta = 1e-05)

cov_exp(x, sq_eta, rho, delta = 1e-05)
}
\arguments{
\item{x}{Distance matrix}

\item{sq_eta}{Parameter for covariance function, see details}

\item{rho}{Parameter for covariance function, see details}

\item{delta}{= 1e-5 Nugget added to diagonal of resulting matrix to keep it
positive definite}

\item{nu05}{Must be an integer in range 0-3. Indicates what type of covariance
function is used. 0 is exponential, 1 is Matérn with nu = 1.5, 2 is Matérn with
nu = 2.5 and 3 is the squared exponential.}
}
\description{
Function for calculating the Matérn/ squared exponential covariance function given
a distance matrix
}
\details{
This function takes a distance matrix and then calculates either the
Matérn kernel with \eqn{\nu} 1/2, 3/2, 5/2, or infinite. Matérn kernel with
infinite nu is equivalent to the squared exponential kernel, and with \eqn{\nu}
= 1/2 the exponential kernel.

Matérn with \eqn{\nu = 1/2}, or exponential: \deqn{sq\_eta * \exp(-x_{ij}/rho)}

Matérn with \eqn{\nu = 3/2}: \deqn{sq\_eta * (1 + (\sqrt{3}x_{ij})/rho) * \exp(-\sqrt{3}x_{ij}/rho)}

Matérn with \eqn{\nu = 5/2}: \deqn{sq\_eta * (1 + \sqrt{5}x_{ij}/rho + 5(x_{ij}/rho)^2 /
  3) * \exp(-\sqrt{5}x_{ij}/rho)}

Matérn with \eqn{\nu = \infty}, or squared exponential: \deqn{sq\_eta *
  \exp(-(x_{ij}^2/2rho^2)}

For more information see the phylogenetic models vignette.
}
\section{Functions}{
\itemize{
\item \code{cov_sq_exp()}: alias for fitting squared exponential

\item \code{cov_exp()}: alias for fitting exponential

}}
