% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{envplot}
\alias{envplot}
\title{Plotting environmental effects on species}
\usage{
envplot(
  object,
  include_intercept = FALSE,
  nrow = NULL,
  y_labels = NULL,
  plotfun = "intervals",
  widths = NULL,
  species = NULL,
  preds = NULL,
  ...
)
}
\arguments{
\item{object}{The jsdmStanFit model object}

\item{include_intercept}{Whether to include the intercept in the plots}

\item{nrow}{The number of rows within the plot}

\item{y_labels}{Which plots should have annotated y axes. Needs to be given
as an integer vector}

\item{plotfun}{Which plot function from mcmc_plot should be used, by default
\code{"intervals"}}

\item{widths}{The widths of the plots}

\item{species}{The species to be included in the plot, by default \code{NULL}
and all species are included. Should be specified as a vector of species
names or numeric indices.}

\item{preds}{The predictors to be included in the plot, by default \code{NULL}
and all predictors are included. Should be specified as a vector of predictor
names or numeric indices.}

\item{...}{Other arguments that are passed to the MCMC plotting function (see
\code{\link[=mcmc_plot.jsdmStanFit]{mcmc_plot.jsdmStanFit()}})}
}
\value{
An object of class \code{"bayesplot_grid"}, for more information see
\code{\link[bayesplot:bayesplot_grid]{bayesplot::bayesplot_grid()}}
}
\description{
Plotting environmental effects on species
}
