% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsdmstanfit-class.R
\name{extract.jsdmStanFit}
\alias{extract.jsdmStanFit}
\alias{extract}
\title{Extract samples from jsdmStanFit object}
\usage{
\method{extract}{jsdmStanFit}(
  object,
  pars,
  permuted = FALSE,
  inc_warmup = FALSE,
  include = TRUE,
  regexp = FALSE
)

extract(object, ...)
}
\arguments{
\item{object}{The jsdmStanFit model object}

\item{pars}{A character vector of parameter names}

\item{permuted}{Whether the draws should be randomly permuted, by default
\code{FALSE}}

\item{inc_warmup}{Whether the warmup period should be included, by default
\code{FALSE}}

\item{include}{Whether the parameters specified by \code{pars} should be included
or excluded from the result, by default \code{TRUE} for inclusion}

\item{regexp}{Whether regular expression matching should be used to match the
contents of \code{pars} to the parameter names, by default \code{FALSE}}
}
\value{
Returns a N by P matrix where N is the number of iterations and P is the
number of parameters. The different chains are appended.
}
\description{
This function extracts named parameters from a jsdmStanFit object, with optional
regular expression matching.
}
\section{Functions}{
\itemize{
\item \code{extract}: Generic method
}}

