% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsdmstanfit-class.R
\name{extract}
\alias{extract}
\alias{extract.default}
\alias{extract.jsdmStanFit}
\title{Extract samples from jsdmStanFit object}
\usage{
extract(object, ...)

\method{extract}{default}(object, ...)

\method{extract}{jsdmStanFit}(object, pars, permuted = FALSE, inc_warmup = FALSE, include = TRUE)
}
\description{
Extract samples from jsdmStanFit object
}
