% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsdmstanfit-class.R
\name{jsdmstan-extractors}
\alias{jsdmstan-extractors}
\alias{get_parnames}
\alias{nuts_params.jsdmStanFit}
\alias{nuts_params}
\alias{log_posterior.jsdmStanFit}
\alias{log_posterior}
\alias{rhat.jsdmStanFit}
\alias{rhat}
\alias{neff_ratio.jsdmStanFit}
\alias{neff_ratio}
\title{Extract quantities useful for model summaries}
\usage{
get_parnames(object, log_lik = FALSE)

\method{nuts_params}{jsdmStanFit}(object, ...)

\method{log_posterior}{jsdmStanFit}(object, ...)

\method{rhat}{jsdmStanFit}(object, ...)

\method{neff_ratio}{jsdmStanFit}(object, ...)
}
\arguments{
\item{object}{The \code{jsdmStanFit} model object}

\item{log_lik}{Whether the log_lik parameters should be included in the parameter
list}

\item{...}{Arguments passed on to the \pkg{bayesplot} equivalent for stanFit
objects}
}
\value{
{
\code{get_parnames()} returns a character vector of model parameter names.

\code{nuts_params()} returns a molten data frame (see \code{\link[reshape2:melt]{reshape2::melt()}}). The
data frame should have columns "Parameter" (factor), "Iteration" (integer),
"Chain" (integer), and "Value" (numeric).

\code{log_posterior()} returns a molten data frame (see \code{\link[reshape2:melt]{reshape2::melt()}}). The
data frame should have columns "Chain" (integer), "Iteration" (integer),
and "Value" (numeric).

\code{rhat()}, \code{neff_ratio()} both return named numeric vectors.
}
}
\description{
These are methdos for extracting various useful summaries from models, including
the model parameter names, NUTS parameters, the log posterior, r-hat and n-eff
ratio.
}
\section{Functions}{
\itemize{
\item \code{get_parnames()}: Get the model parameter names

\item \code{nuts_params(jsdmStanFit)}: Get the NUTS parameters

\item \code{log_posterior(jsdmStanFit)}: Get the log posterior

\item \code{rhat(jsdmStanFit)}: Get the R-hat

\item \code{neff_ratio(jsdmStanFit)}: Get the n_eff ratio

}}
