% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsdmstan-package.R
\docType{package}
\name{jsdmstan-package}
\alias{jsdmstan-package}
\alias{jsdmstan}
\title{The 'jsdmstan' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.26.1. https://mc-stan.org
}
