% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsdmstan-families.R
\name{jsdmStanFamily}
\alias{jsdmStanFamily}
\alias{jsdmStanFamily_empty}
\title{jsdmStanFamily class}
\usage{
jsdmStanFamily_empty()
}
\description{
This is the jsdmStanFamily class, which occupies a slot within any
jsdmStanFit object.
}
\section{Elements for \code{jsdmStanFamily} objects}{

\describe{
\item{\code{family}}{
A length one character vector describing family used to fit object. Options
are \code{"gaussian"}, \code{"poisson"}, \code{"bernoulli"},
\code{"neg_binomial"}, \code{"binomial"}, \code{"zi_poisson"},
\code{"zi_neg_binomial"}, or \code{"multiple"}.
}
\item{\code{params}}{
A character vector that includes all the names of the family-specific parameters.
}
\item{\code{params_dataresp}}{
A character vector that includes any named family-specific parameters that are
modelled in response to data.
}
\item{\code{preds}}{
A character vector of the measured predictors included if family parameters
are modelled in response to data. If family parameters are not modelled in
response to data this is left empty.
}
\item{\code{data_list}}{
A list containing the original data used to fit the model
(empty when save_data is set to \code{FALSE} or family parameters are not
modelled in response to data).
}
}
}

