% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsdmstanfit-class.R
\name{jsdmStanFit}
\alias{jsdmStanFit}
\alias{print.jsdmStanFit}
\alias{extract}
\title{jsdmStanFit class}
\usage{
\method{print}{jsdmStanFit}(object)

extract(object, pars, permuted = FALSE, inc_warmup = FALSE, include = TRUE)
}
\description{
This is the jsdmStanFit class, which stan_gllvm and stan_mglmm both create.
}
\section{Functions}{
\itemize{
\item \code{print.jsdmStanFit}: Print the default summary for the model

\item \code{extract}: Wrapper for extracting samples from model fit
}}

\section{Elements for \code{jsdmStanFit} objects}{

\describe{
\item{\code{jsdm_type}}{
A length one character vector describing type of jSDM
}
\item{\code{family}}{
A character vector describing response family
}
\item{\code{species}}{
A character vector of the species names
}
\item{\code{sites}}{
character vector of the site IDs
}
\item{\code{preds}}{
A character vector of the measured predictors included
}
\item{\code{data_list}}{
A list containing the original data used to fit the model
(empty when save_data is set to FALSE)
}
\item{\code{n_latent}}{
A length one integer vector representing number of latent
variables (in gllvm type fits) or NA in all other cases
}
}
}

