% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{loo.jsdmStanFit}
\alias{loo.jsdmStanFit}
\alias{loo}
\title{Efficient approximate leave-one-out cross-validation using the \pkg{loo} package}
\usage{
\method{loo}{jsdmStanFit}(x, calc_reff = TRUE, cores = getOption("mc.cores", 1), ...)
}
\arguments{
\item{x}{The jsdmStanFit model object}

\item{calc_reff}{Whether to calculate the relative efficiencies for loo, by
default \code{TRUE}. If set to \code{FALSE} then relative efficiency is
assumed to be 1.}

\item{cores}{The number of cores the loo functions use, by default uses the
mc.cores option (or 1, if unspecified).}

\item{...}{Other arguments passed to the \code{\link[loo]{loo}} function}
}
\value{
A list with class \code{c("psis_loo","loo")}, as detailed in the
\code{\link[loo]{loo}} documentation
}
\description{
This function uses the \pkg{loo} package to compute PSIS-LOO CV, efficient
approximate leave-one-out (LOO) cross-validation for Bayesian models using Pareto
smoothed importance sampling (PSIS). This requires that the model was fit using
\code{log_lik = TRUE}.
}
