% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_jsdm.R
\name{make_smooth}
\alias{make_smooth}
\title{Generate Stan code for a smoother from mgcv}
\usage{
make_smooth(sterms, data)
}
\arguments{
\item{sterms}{a smooth term as generated by [\link[mgcv:interpret.gam]{mgcv::interpret.gam}]}

\item{data}{a \code{data.frame} that contains all the variables in \code{sterms}}
}
\value{
a \code{list} of data to give to Stan
}
\description{
This function creates Stan code for an \code{mgcv} smoother. It is (roughly)
based on \link[mgcv:jagam]{mgcv::jagam}.
}
\author{
David L Miller
}
\keyword{internal}
