% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_check.R
\name{multi_pp_check}
\alias{multi_pp_check}
\title{Multiple pp_check plots per species}
\usage{
multi_pp_check(
  object,
  plotfun = "dens_overlay",
  species = NULL,
  ndraws = NULL,
  xlim = NULL,
  ylim = NULL,
  titles = character(),
  subtitles = character(),
  legends = TRUE,
  save_gg_objects = TRUE,
  grid_args = list(...),
  ...
)
}
\arguments{
\item{object}{The jsdmStanFit model object}

\item{plotfun}{The ppc plot function to use, given as a character string. The
default is to call \link[bayesplot:PPC-distributions]{ppc_dens_overlay}. Can be
specified as either the entire name of function as a character string or
without the ppc_ prefix.}

\item{species}{Which species should be included, by default all}

\item{ndraws}{How many draws should be used within the plots}

\item{xlim, ylim}{Optionally, numeric vectors of length 2 specifying lower and
upper limits for the axes that will be shared across all plots.}

\item{titles, subtitles}{Optional character vectors of plot titles and
subtitles. If specified, \code{titles} and \code{subtitles} must must have
length equal to the number of plots specified.}

\item{legends}{If any of the plots have legends should they be displayed?
Defaults to \code{TRUE}.}

\item{save_gg_objects}{If \code{TRUE}, the default, then the ggplot objects
specified in \code{...} or via the \code{plots} argument are saved in a
list in the \code{"bayesplots"} component of the returned object.
Setting this to \code{FALSE} will make the returned object smaller but
these individual plot objects will not be available.}

\item{grid_args}{An optional named list of arguments to pass to
\code{\link[gridExtra:arrangeGrob]{gridExtra::arrangeGrob()}} (\code{nrow}, \code{ncol},
\code{widths}, etc.).}

\item{...}{Other options passed to the bayesplot PPC function or
posterior_predict. Note that the same value will be used for every species
included (e.g. binwidth for a histogram cannot be specified by species).}
}
\value{
An object of class \code{"bayesplot_grid"}, for more information see
\code{\link[bayesplot:bayesplot_grid]{bayesplot::bayesplot_grid()}}
}
\description{
Multiple pp_check plots per species
}
