% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ordiplot}
\alias{ordiplot}
\title{Plotting an ordination plot for a GLLVM model}
\usage{
ordiplot(
  object,
  choices = c(1, 2),
  type = "species",
  summary_stat = "mean",
  ndraws = 20,
  draw_ids = NULL,
  size = c(2, 1),
  alpha = c(1, 0.5),
  shape = c(18, 16)
)
}
\arguments{
\item{object}{The \code{jsdmStanFit} model object}

\item{choices}{Which latent variables to plot as dimensions, by default
\code{c(1,2)}}

\item{type}{Whether to plot sites or species, default \code{"species"}.}

\item{summary_stat}{The summary statistic used to plot overall averages of the
posterior sample. By default this is \code{"mean"}, and \code{NULL} will result
in no summary being included}

\item{ndraws}{How many individual draws to include in plot, by default \code{20}. Setting
this to \code{0} will result in no individual draws being included}

\item{draw_ids}{Which draws to include in plot (overrides \code{ndraws})}

\item{size}{The size of the points in the graph, specified as a two-element vector
with the first being used for the summary points and the second the individual
draws, default \code{c(2,1)}}

\item{alpha}{The transparency/alpha of the points in the graph, specified as a
two-element vector with the first being used for the summary points and the
second the individual draws, default \code{c(1,0.5)}}

\item{shape}{The shape of the points in the graph, specified as a two-element
vector with the first being used for the summary points and the second the
individual draws, default \code{c(18,16)}}
}
\value{
A \link[ggplot2:ggplot]{ggplot} object that can be customised using the
\pkg{ggplot2} package
}
\description{
This function takes a GLLVM model fit and plots an ordination plot with a random
(or specified) selection of draws
}
\examples{

\dontrun{
 # First simulate data and fit the model:
 gllvm_data <- jsdm_sim_data(method = "gllvm", N = 100, S = 6, D = 3,
                             family = "bernoulli")
 gllvm_fit <- stan_jsdm(dat_list = gllvm_data, method = "gllvm",
                        family = "bernoulli")

ordiplot(gllvm_fit)
# now plot the 1st and 3rd latent variables against each other for the sites:
ordiplot(gllvm_fit, choices = c(1,3), type = "sites")

}
}
