% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinewood-data.R
\docType{data}
\name{pinewood_env}
\alias{pinewood_env}
\title{Environmental data from native pinewoods in Scotland, 1971}
\format{
A data frame with 25 rows and 7 columns.
\describe{
\item{Site_no}{The site identifier (27 in total)}
\item{OSGR}{The 1 km grid reference of each site (OSGB)}
\item{rainfall}{The average annual rainfall in mm, divided by 1000 and minus 1.5}
\item{tasmax_summer}{The average maximum temperature in summer in deg Celsius, minus 16}
\item{tasmin_winter}{The average minimum temperature in winter in deg Celsius, plus 1}
\item{Soil_PH}{The soil pH of the top 10cm of soil, minus 4.3}
\item{No_plots}{The difference between the number of plots surveyed and 16 (maximum number of plots surveyed)}
}
}
\source{
\url{https://doi.org/10.5285/56a48373-771c-4d4a-8b5a-45ef496c6e55}
}
\usage{
pinewood_env
}
\description{
This is the environmental data from a series of plots across native Scots Pine
woodlands in Scotland surveyed in 1971. It contains the approximate location of
each site, the soil pH of each site plus the average annual rainfall, maximum summer
temperature and minimum winter temperature from 1941-1970. All climate variables
were calculated based upon the HadUK-Grid climate data (Met Office et al., 2022).
For more information on the sites see Wood and Bunce (2016). All climate and soils
data have been scaled such that they are centred approximately around zero and
have a range within single digits.
}
\details{
This data is made available under the terms of the Open Government Licence.
}
\references{
{
Wood, C. M., & Bunce, R. G. H. (2016). Ecological survey of the native pinewoods
of Scotland 1971. Earth System Science Data, 8(1), 177-189.
https://doi.org/10.5194/essd-8-177-2016

Met Office; Hollis, D.; McCarthy, M.; Kendon, M.; Legg, T. (2022): HadUK-Grid
Gridded Climate Observations on a 1km grid over the UK, v1.1.0.0 (1836-2021). NERC
EDS Centre for Environmental Data Analysis, 26 May 2022.
http://dx.doi.org/10.5285/bbca3267dc7d4219af484976734c9527
}
}
\seealso{
\code{\link{pinewood}}
}
\keyword{datasets}
