% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.jsdmStanFit}
\alias{plot.jsdmStanFit}
\title{Plot the traceplots and density plots for parameters within a jsdmStanFit object}
\usage{
\method{plot}{jsdmStanFit}(
  x,
  pars = NULL,
  combo = c("dens", "trace"),
  N = 5L,
  ask = TRUE,
  inc_warmup = FALSE,
  include = TRUE,
  sample_n = 10,
  regexp = FALSE,
  plot = TRUE,
  newpage = TRUE,
  ...
)
}
\arguments{
\item{x}{The \code{jsdmStanFit} model object}

\item{pars}{The parameters to plot, by default a random sample of twenty of the
parameters fit within the model}

\item{combo}{Which combination of plot types within \code{bayesplot::mcmc_combo}
to use, by default \code{c("dens", "trace")}}

\item{N}{The number of plots per page, default \code{5}}

\item{ask}{Whether to ask before plotting a new page, default \code{TRUE}}

\item{inc_warmup}{Whether to include the warmup period in the plots, by default
\code{FALSE}}

\item{include}{Whether to include or exclude the parameters specified by pars, by
default \code{TRUE} (i.e. include)}

\item{sample_n}{If \code{pars = NULL} then the number of random non-sigma
parameters to include (details in description)}

\item{regexp}{If pars should be treated as a regular expression for matching to
parnames, by default \code{FALSE}}

\item{plot}{Whether to plot the plots, default \code{TRUE}}

\item{newpage}{Whether the first plot should be plotted on a new page, by default
\code{TRUE}}

\item{...}{Arguments passed to \code{bayesplot::mcmc_combo}}
}
\value{
An invisible list of the plots
}
\description{
This function takes parameters from the jsdmStanfit object and plots them using
the \code{mcmc_combo} function from the \code{bayesplot} package. These models
have a lot of parameters so selecting a subset is recommended. If pars is set to
\code{NULL} (the default) all parameters with either sigma or kappa in their name
will be plotted along with a random selection of the other parameters (total
number of other parameters set by \code{sample_n}).
}
