% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predict.R
\name{posterior_predict.jsdmStanFit}
\alias{posterior_predict.jsdmStanFit}
\alias{posterior_predict}
\title{Draw from the posterior predictive distribution}
\usage{
\method{posterior_predict}{jsdmStanFit}(
  object,
  newdata = NULL,
  newdata_type = "X",
  ndraws = NULL,
  draw_ids = NULL,
  list_index = "draws",
  Ntrials = NULL,
  ...
)
}
\arguments{
\item{object}{The model object}

\item{newdata}{New data, by default \code{NULL} and uses original data}

\item{newdata_type}{What form is the new data in, at the moment only
supplying covariates is supported.}

\item{ndraws}{Number of draws, by default the number of samples in the
posterior. Will be sampled randomly from the chains if fewer than the
number of samples.}

\item{draw_ids}{The IDs of the draws to be used, as a numeric vector}

\item{list_index}{Whether to return the output list indexed by the number of
draws (default), species, or site.}

\item{Ntrials}{For the binomial distribution the number of trials, given as
either a single integer which is assumed to be constant across sites or as
a site-length vector of integers.}

\item{...}{Currently unused}
}
\value{
A list of linear predictors. If list_index is \code{"draws"} (the default)
the list will have length equal to the number of draws with each element of
the list being a site x species matrix. If the list_index is \code{"species"} the
list will have length equal to the number of species with each element of
the list being a draws x sites matrix. If the list_index is \code{"sites"} the
list will have length equal to the number of sites with each element of
the list being a draws x species matrix.
}
\description{
Draw from the posterior predictive distribution of the outcome.
}
\seealso{
\code{\link[=posterior_linpred.jsdmStanFit]{posterior_linpred.jsdmStanFit()}}
}
