% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsdmstan-families.R
\name{print.jsdmStanFamily}
\alias{print.jsdmStanFamily}
\title{Print jsdmStanFamily object}
\usage{
\method{print}{jsdmStanFamily}(x, ...)
}
\arguments{
\item{x}{A jsdmStanFamily object}

\item{...}{Other arguments, not used at this stage.}
}
\description{
Print jsdmStanFamily object
}
