% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsdmstanfit-class.R
\name{print.jsdmStanFit}
\alias{print.jsdmStanFit}
\title{Print the default summary for the model}
\usage{
\method{print}{jsdmStanFit}(x, ...)
}
\arguments{
\item{x}{The \code{jsdmStanFit} model object}

\item{...}{Other arguments passed to \link{summary.jsdmStanFit}}
}
\description{
This prints out a summary for the models which includes the type of model fit, the
number of species, sites and predictors as well as a summary of any parameters
with Rhat > 1.01 or effective sample size to total number of samples ratio < 0.05
}
