% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data_funs.R
\name{sim_helpers}
\alias{sim_helpers}
\alias{rgampois}
\alias{rlkj}
\title{Helper functions for simulating data}
\usage{
rgampois(n, mu, scale)

rlkj(n, eta = 1, cholesky = FALSE)
}
\arguments{
\item{n}{The number of samples to create/dimension of correlation matrix}

\item{mu}{The mean of the negative binomial parameterisation}

\item{scale}{The phi parameter that controls overdispersion of the negative
binomial distribution (see details for description)}

\item{eta}{The shape parameter of the LKJ distribution}

\item{cholesky}{Whether the correlation matrix should be returned as the Cholesky
decomposition, by default \code{FALSE}}
}
\description{
The \code{rlkj} function is for generating random LKJ correlation matrices and the
\code{rgampois} function generates random draws from the Stan's alternative
parameterisation of the negative binomial distribution.
}
\details{
The Lewandowski-Kurowicka-Joe (LKJ) distribution is a prior distribution for
correlation matrices, with the shape parameter eta. If eta is 1 then the density
is uniform over the correlation matrix, ith eta > 1 then the the probability
concentrates around the identity matrix while is 0 < eta < 1 the probability
concentrates away from the identity matrix.

The alternative parameterisation of the negative binomial distribution is:

\deqn{NegBinomial2(y | mu, scale) = binom(y+scale-1,y) (mu/mu+scale)^y (scale/mu + scale)^scale}

Where the mean of the distribution is mu and the variance is \eqn{mu + (mu^2/scale)}

The \code{rlkj} function are sourced from Ben Goodrich's response on the Stan
google mailing list. (see link
\url{https://groups.google.com/g/stan-users/c/3gDvAs_qwN8/m/Xpgi2rPlx68J)}). The
\code{rgampois} function is sourced from the rethinking package by Richard McElreath.
}
