% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{smoothplot}
\alias{smoothplot}
\title{Plot the smooth effect of any spline within a model}
\usage{
smoothplot(object, ndraws = 20, alpha = 0.4, draw_ids = NULL, summarise = NULL)
}
\arguments{
\item{object}{The jsdmStanFit model object}

\item{ndraws}{The number of draws of the predictor}

\item{alpha}{The alpha (transparency) value of the lines}

\item{draw_ids}{Which draws to include in plot (overrides \code{ndraws})}

\item{summarise}{Whether to include a summary, by default \code{NULL} i.e. no
summary. Can also be given as the name of a function (e.g. \code{"mean"}).
Note that this summarisation only occurs over the number of draws
specified.}
}
\value{
a list of ggplot objects
}
\description{
Plot the smooth effect of any spline within a model
}
