% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_mglmm.R
\name{stan_mglmm}
\alias{stan_mglmm}
\title{Run mglmm models in Stan}
\usage{
stan_mglmm(Y = NULL, X = NULL, intercept = TRUE, dat_list = NULL, family, ...)
}
\arguments{
\item{Y}{Matrix of species by sites. Rows are assumed to be sites, columns are
assumed to be species}

\item{X}{The covariates matrix, with rows being site and columns being covariates}

\item{intercept}{Whether the model should be fit with an intercept}

\item{family}{The response family for the model, required to be one of "gaussian",
"bernoulli", "poisson" or "neg_binomial"}

\item{...}{Arguments passed to rstan::sampling}
}
\description{
Run mglmm models in Stan
}
\details{

}
