% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance_funs.R
\name{cov_matern}
\alias{cov_matern}
\alias{cov_sq_exp}
\title{Calculate covariance functio based on a given data matrix}
\usage{
cov_matern(x, sq_eta, rho, delta = 1e-05, nu05)

cov_sq_exp(x, sq_eta, rho, delta = 1e-05)
}
\arguments{
\item{x}{Distance matrix}

\item{sq_eta}{Parameter for covariance function, see details}

\item{rho}{Parameter for covariance function, see details}

\item{delta}{= 1e-5 Nugget added to diagonal of resulting matrix to keep it
positive definite}

\item{nu05}{Must be an integer in range 0-3. Indicates what type of covariance
function is used. 0 is exponential, 1 is Matérn with nu = 1.5, 2 is Matérn with
nu = 2.5 and 3 is the squared exponential.}
}
\description{
Function for calculating the Matérn/ squared exponential covariance function given
a distance matrix
}
\details{
This function takes a distance matrix and then calculates either the
Matérn kernel with \eqn{\nu} 1/2, 3/2, 5/2, or infinite. Matérn kernel with
infinite nu is equivalent to the squared exponential kernel, and with \eqn{\nu}
= 1/2 the exponential kernel.

Matérn with \eqn{\nu = 1/2} (i.e. exponential): \deqn{sq_eta * exp(-x[i,j]/rho)}

Matérn with \eqn{\nu = 3/2}: \deqn{sq_eta * (1 + x[i,j]/rho) * exp(-x[i,j]/rho)}

Matérn with \eqn{\nu = 5/2}: \deqn{sq_eta * (1 + x[i,j]/rho + (x[i,j]/rho)^2 /
  3) * exp(-x[i,j]/rho)}

Matérn with \eqn{\nu = Inf}, or squared exponential: \deqn{sq_eta *
  exp(-(x[i,j]/rho)^2)/2}
}
\section{Functions}{
\itemize{
\item \code{cov_sq_exp}: alias for fitting squared exponential with
}}

