% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsdmstanfit-class.R
\name{extract.jsdmStanFit}
\alias{extract.jsdmStanFit}
\alias{extract}
\title{Extract samples from jsdmStanFit object}
\usage{
\method{extract}{jsdmStanFit}(
  object,
  pars = NULL,
  permuted = FALSE,
  inc_warmup = FALSE,
  include = TRUE,
  regexp = FALSE,
  return_array = FALSE,
  ...
)

extract(object, ...)
}
\arguments{
\item{object}{The jsdmStanFit model object}

\item{pars}{A character vector of parameter names}

\item{permuted}{Whether the draws should be randomly permuted, by default
\code{FALSE}}

\item{inc_warmup}{Whether the warmup period should be included, by default
\code{FALSE}}

\item{include}{Whether the parameters specified by \code{pars} should be included
or excluded from the result, by default \code{TRUE} for inclusion}

\item{regexp}{Whether regular expression matching should be used to match the
contents of \code{pars} to the parameter names, by default \code{FALSE}}

\item{return_array}{Whether to return the output as a 3 dimensional array
(\code{TRUE}) or a named list (\code{FALSE}, the default)}

\item{...}{Arguments passed to \code{\link[=get_parnames]{get_parnames()}}}
}
\value{
If \code{return_array = FALSE} returns a named list with each parameter
group being an element of the list. Each list element is an array with the first
dimension being the iteration (all chains are appended) and the other dimensions
coming from the parameter dimensions. If \code{return_array = TRUE} then a 3
dimensional array is returned with the first dimension being the iterations, the
second the chains and the third the parameters.
}
\description{
This function extracts named parameters from a jsdmStanFit object, with optional
regular expression matching.
}
\section{Functions}{
\itemize{
\item \code{extract}: Generic method
}}

