% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{jsdm_prior}
\alias{jsdm_prior}
\alias{print.jsdmprior}
\title{Create prior object for jsdmstan models and data simulation}
\usage{
jsdm_prior(
  sigmas_preds = "normal(0,1)",
  z_preds = "normal(0,1)",
  cor_preds = "lkj_corr_cholesky(1)",
  a = "normal(0,1)",
  a_bar = "normal(0,1)",
  sigma_a = "normal(0,1)",
  sigmas_species = "normal(0,1)",
  z_species = "normal(0,1)",
  cor_species = "lkj_corr_cholesky(1)",
  LV = "normal(0,1)",
  L = "normal(0,1)",
  sigma_L = "normal(0,1)",
  sigma = "normal(0,1)",
  kappa = "normal(0,1)"
)

\method{print}{jsdmprior}(x, ...)
}
\arguments{
\item{sigmas_preds}{The standard deviation of the covariate effects, constrained to be
positive (default standard normal)}

\item{z_preds}{The covariate effects (default standard normal)}

\item{cor_preds}{The correlation matrix on the covariate effects (npred by npred
matrix represented as a Cholesky factor of a correlation matrix) (default
\code{"lkj_corr_cholesky(1)"})}

\item{a}{The site level intercepts (default standard normal)}

\item{a_bar}{The mean site level intercept}

\item{sigma_a}{The standard deviation of the site level intercepts, constrained to
be positive and default prior is half standard normal}

\item{sigmas_species}{For MGLMM method, the standard deviations of the species
covariances, constrained to be positive (default half standard normal)}

\item{z_species}{For MGLMM method, the S by N matrix of species covariance by site
(default standard normal)}

\item{cor_species}{For MGLMM method, the correlation between species represented
as a Cholesky factor correlation matrix (default \code{"lkj_corr_cholesky(1)"})}

\item{LV}{For GLLVM method, the per site latent variable loadings (default
standard normal)}

\item{L}{For GLLVM method, the non-zero species latent variable loadings (default
standard normal)}

\item{sigma_L}{For GLLVM method, the variance of the species loadings, constrained
to be positive (default standard normal)}

\item{sigma}{For Gaussian response, the standard deviation parameter. Constrained
to be positive (default standard normal)}

\item{kappa}{For negative binomial response, the negative binomial variance
parameter. Constrained to be positive (default standard normal)}

\item{x}{Object of class \code{jsdmprior}}

\item{...}{Currently unused}
}
\value{
An object of class \code{"jsdmprior"} taking the form of a named list
}
\description{
This function creates all of the potential priors used within a \code{jsdmstan}
model and can be used as the input to all \code{\link[=stan_jsdm]{stan_jsdm()}} family of functions and
the \code{\link[=jsdm_sim_data]{jsdm_sim_data()}} functions.
}
\details{
Each prior has to be specified as a character string corresponding to the
appropriate stan command. The most common versions of these are supported by the
simulated data functions, however there are functions that can be fed to the stan
fitting procedure that will not be able to be used as input for \code{\link[=jsdm_sim_data]{jsdm_sim_data()}}.
Parameters \code{sigmas_preds}, \code{sigma_a}, \code{sigmas_species}, \code{sigma_L},
\code{sigma}, and \code{kappa} are fixed to be positive only in the stan code and
this cannot be changed. Parameters \code{cor_preds} and \code{cor_species} are
assumed to be the Cholesky factor of a correlation matrix. All other parameters
are real numbers. For all parameters that represent vectors or matrices the prior
has to be the same across the entire vector or matrix (note that for the species
latent variable loadings in a GLLVM model the prior is set on the non-zero matrix
components \code{L} and not on the entire matrix).

Prior distributions supported by \code{\link[=jsdm_sim_data]{jsdm_sim_data()}} are \code{"normal(mean, sd)"},
\code{"student_t(df, mu, sigma)"}, \code{"cauchy(location, scale)"},
\code{"gamma(shape, scale)"}, \code{"inv_gamma(shape, scale)"} and
\code{"lkj_corr_cholesky(eta)"}.
}
\section{Functions}{
\itemize{
\item \code{print(jsdmprior)}: Print method for object of class \code{jsdmprior}

}}
\examples{
pr <- jsdm_prior(kappa = "gamma(0.01,0.01)")
pr

}
\seealso{
\link{sim_helpers} for a description of the parameterisations used within the
data simulation functions
}
