% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data_funs.R
\name{jsdm_sim_data}
\alias{jsdm_sim_data}
\alias{gllvm_sim_data}
\alias{mglmm_sim_data}
\title{Generate simulated data within a variety of jSDM methodologies}
\usage{
jsdm_sim_data(
  N,
  S,
  D = NULL,
  K = 0L,
  family,
  method = c("gllvm", "mglmm"),
  species_intercept = TRUE,
  site_intercept = FALSE,
  prior = jsdm_prior()
)

gllvm_sim_data(...)

mglmm_sim_data(...)
}
\arguments{
\item{N}{is number of sites}

\item{S}{is number of species}

\item{D}{is number of latent variables, used within gllvm method}

\item{K}{is number of covariates, by default \code{0}}

\item{family}{is the response family, must be one of \code{"gaussian"},
\code{"neg_binomial"}, \code{"poisson"} or \code{"bernoulli"}. Regular expression
matching is supported.}

\item{method}{is the jSDM method to use, currently either \code{"gllvm"} or
\code{"mglmm"} - see details for more information.}

\item{species_intercept}{Whether to include an intercept in the predictors, must be
\code{TRUE} if \code{K} is \code{0}. Defaults to \code{TRUE}.}

\item{site_intercept}{Whether to include a site intercept. Defaults to \code{FALSE}.}

\item{prior}{Set of prior specifications from call to \code{\link[=jsdm_prior]{jsdm_prior()}}}

\item{...}{Arguments passed to jsdm_sim_data}
}
\description{
The \code{jsdm_sim_data} function can simulate data with either a multivariate
generalised mixed model (MGLMM) or a generalised linear latent variable model
(GLLVM). The \code{gllvm_sim_data} and \code{mglmm_sim_data} are aliases for
\code{jsdm_sim_data} that set \code{method} to \code{"gllvm"} and \code{"mglmm"}
respectively.
}
\details{
This simulates data based on a joint species distribution model with
either a generalised linear latent variable model approach or a multivariate
generalised linear mixed model approach.

Models can be fit with or without "measured predictors", and if measured
predictors are included then the species have species-specific parameter
estimates. These can either be simulated completely independently, or have
information pooled across species. If information is pooled this can be modelled
as either a random draw from some mean and standard deviation or species
covariance can be modelled together (this will be the covariance used in the
overall model if the method used has covariance).
}
\section{Functions}{
\itemize{
\item \code{gllvm_sim_data()}: Alias for \code{jsdm_sim_data} with \code{method =
"gllvm"}

\item \code{mglmm_sim_data()}: Alias for \code{jsdm_sim_data} with \code{method =
"mglmm"}

}}
