% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data_funs.R
\name{jsdm_sim_data}
\alias{jsdm_sim_data}
\alias{gllvm_sim_data}
\alias{mglmm_sim_data}
\title{Generate simulated data within a variety of jSDM methodologies}
\usage{
jsdm_sim_data(
  N,
  S,
  D = NULL,
  K = 0L,
  family,
  method = c("gllvm", "mglmm"),
  phylo = FALSE,
  species_intercept = TRUE,
  site_intercept = FALSE,
  delta = NULL,
  nu05 = NULL,
  eta = 1
)

gllvm_sim_data(...)

mglmm_sim_data(...)
}
\arguments{
\item{N}{is number of sites}

\item{S}{is number of species}

\item{D}{is number of latent variables, used within gllvm method}

\item{K}{is number of covariates, by default 0}

\item{family}{is the response family, must be one of "gaussian", "neg_binomial",
"poisson" or "bernoulli"}

\item{method}{is the jSDM method to use, currently either GLLVM or MGLMM - see
details for more information.}

\item{phylo}{is whether to randomly generate a phylogenetic tree that will be used
to constrain beta estimates. Can be given as TRUE, FALSE or as a given distance
matrix.}

\item{species_intercept}{Whether to include an intercept in the predictors, must be
TRUE if K is 0. Defaults to TRUE.}

\item{site_intercept}{Whether to include a site intercept. Defaults to FALSE.}

\item{delta}{Nugget added to diagonal of resulting matrix to keep it positive
definite}

\item{nu05}{Must be an integer in range 0-3. Indicates what type of covariance
function is used. 0 is exponential, 1 is Matérn with nu = 1.5, 2 is Matérn with
nu = 2.5 and 3 is the squared exponential.}

\item{eta}{Shape parameter in random generation of LKJ matrix, defaults to 1.}

\item{...}{Arguments passed to jsdm_sim_data}
}
\description{
The \code{jsdm_sim_data} function can simulate data with either a multivariate
generalised mixed model (MGLMM) or a generalised linear latent variable model
(GLLVM). The \code{gllvm_sim_data} and \code{mglmm_sim_data} are aliases for
\code{jsdm_sim_data} that set \code{method} to \code{"gllvm"} and \code{"mglmm"}
respectively.
}
\details{
This simulates data based on a joint species distribution model with
either a generalised linear latent variable model approach or a multivariate
generalised linear mixed model approach.

Models can be fit with or without "measured predictors", and if measured
predictors are included then the species have species-specific parameter
estimates. These can either be simulated completely independently, or have
information pooled across species. If information is pooled this can be modelled
as either a random draw from some mean and standard deviation or species
covariance can be modelled together (this will be the covariance used in the
overall model if the method used has covariance).

It makes a whole load of assumptions, incl that all parameters and measured
predictors are normally distributed with mean 0 and standard deviation 1, with
variance parameters being the absolute of this distribution. The exceptions are
that if there is a Matern (or exponential quadratic) kernel being fit the etasq
and rho parameters are either simulated as an inverse gamma with shape 10 and
scale 0.1 if the invgamma package is installed or as an absolute value of a
normal distribution with mean 1 and standard deviation 0.2. If no phylogenetic
matrix is supplied the correlations between species are simulated as a random
draw from a LKJ correlation matrix with eta = 1.
}
\section{Functions}{
\itemize{
\item \code{gllvm_sim_data}: Alias for \code{jsdm_sim_data} with \code{method =
"gllvm"}

\item \code{mglmm_sim_data}: Alias for \code{jsdm_sim_data} with \code{method =
"mglmm"}
}}

