% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsdm_stancode.R
\name{jsdm_stancode}
\alias{jsdm_stancode}
\alias{print.jsdmstan_model}
\title{Make stancode for the jsdm model}
\usage{
jsdm_stancode(
  method,
  family,
  prior = jsdm_prior(),
  log_lik = TRUE,
  site_intercept = "none"
)

\method{print}{jsdmstan_model}(x, ...)
}
\arguments{
\item{method}{The method, one of \code{"gllvm"} or \code{"mglmm"}}

\item{family}{The family, one of "\code{"gaussian"}, \code{"bernoulli"},
\code{"poisson"} or \code{"neg_binomial"}}

\item{prior}{The prior, given as the result of a call to \code{\link[=jsdm_prior]{jsdm_prior()}}}

\item{log_lik}{Whether the log likelihood should be calculated in the generated
quantities (by default \code{TRUE}), required for loo}

\item{site_intercept}{Whether a site intercept should be included, potential
values \code{"none"} (no site intercept), \code{"grouped"} (a site intercept
with hierarchical grouping) or \code{"ungrouped"} (site intercept with no
grouping)}

\item{x}{The jsdm_stancode object}

\item{...}{Currently unused}
}
\value{
A character vector of Stan code, class "jsdmstan_model"
}
\description{
This function returns the Stan code used to fit the model as specified by the data
list, family and method.
}
\section{Functions}{
\itemize{
\item \code{print(jsdmstan_model)}: A printing function for jsdmstan_model objects

}}
\examples{
jsdm_stancode(family = "gaussian", method = "gllvm")
jsdm_stancode(family = "poisson", method = "mglmm")

}
