% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_check.R
\name{jsdm_statsummary}
\alias{jsdm_statsummary}
\title{Extract summary statistics for a \code{jsdmStanFit} model}
\usage{
jsdm_statsummary(
  object,
  species = NULL,
  sites = NULL,
  summary_stat = "sum",
  post_type = "linpred",
  calc_over = "site",
  simplify = TRUE,
  ...
)
}
\arguments{
\item{object}{A \code{jsdmStanFit} model object}

\item{species}{Which species to include in the summary statistic, by default all}

\item{sites}{Which sites to include in the summary statistic, by default all}

\item{summary_stat}{The summary statistic to be used, by default \code{sum} but any
function can be used.}

\item{post_type}{The type of posterior prediction to be used, either
\code{"linpred"} for \link[jsdmstan]{posterior_linpred} or \code{"predict"} for
\link[jsdmstan]{posterior_predict}}

\item{calc_over}{Whether to calculate the summary statistic by site or species, by
default \code{species}}

\item{simplify}{Whether to simplify the output into a matrix, by default
\code{TRUE}}

\item{...}{Arguments passed to the posterior prediction function}
}
\value{

}
\description{
Extract summary statistics for a \code{jsdmStanFit} model
}
