% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsdmstan-package.R
\docType{package}
\name{jsdmstan-package}
\alias{jsdmstan-package}
\alias{jsdmstan}
\title{The 'jsdmstan' package.}
\description{
A package for fitting joint species distribution models (jSDMs) in
Stan.

This package can be used to both simulate and fit data according to a
multivariate generalised linear mixed model (MGLMM) or a generalised linear
latent variable model (GLLVM). All models are fit using the \pkg{rstan} package.
Summary functions are provided, as are interfaces to the \pkg{bayesplot}
plotting functions
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package
version 2.26.1. https://mc-stan.org
}
