% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsdmstanfit-class.R
\docType{class}
\name{jsdmStanFit-class}
\alias{jsdmStanFit-class}
\title{jsdmStanFit fitted jsdm_stan model}
\description{
jsdmStanFit fitted jsdm_stan model
}
\section{Slots}{

\describe{
\item{\code{jsdm_type}}{A length one character vector describing type of jSDM}

\item{\code{n_species}}{A length one integer vector representing number of species}

\item{\code{n_site}}{A length one integer vector representing number of sites}

\item{\code{n_preds}}{A length two integer vector representing number of preds then
whether intercept included}

\item{\code{n_latent}}{A length one integer vector representing number of latent
variables (in gllvm type fits) or NA in all other cases}
}}

