% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsdmstanfit-class.R
\name{jsdmStanFit}
\alias{jsdmStanFit}
\title{jsdmStanFit class}
\description{
This is the jsdmStanFit class, which stan_gllvm and stan_mglmm both create.
}
\section{Elements for \code{jsdmStanFit} objects}{

\describe{
\item{\code{jsdm_type}}{
A length one character vector describing type of jSDM
}
\item{\code{family}}{
A jsdmStanFamily object describing characteristics of family
}
\item{\code{species}}{
A character vector of the species names
}
\item{\code{sites}}{
character vector of the site IDs
}
\item{\code{preds}}{
A character vector of the measured predictors included
}
\item{\code{data_list}}{
A list containing the original data used to fit the model
(empty when save_data is set to \code{FALSE})
}
\item{\code{n_latent}}{
A length one integer vector representing number of latent
variables (in gllvm type fits) or \code{NA} in all other cases
}
}
}

