% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsdmstanfit-class.R
\name{jsdmStanFit}
\alias{jsdmStanFit}
\alias{jsdmStanFit_empty}
\alias{print.jsdmStanFit}
\alias{get_parnames}
\alias{nuts_params.jsdmStanFit}
\alias{log_posterior.jsdmStanFit}
\alias{rhat.jsdmStanFit}
\alias{neff_ratio.jsdmStanFit}
\title{jsdmStanFit class}
\usage{
jsdmStanFit_empty()

\method{print}{jsdmStanFit}(object)

get_parnames(x)

\method{nuts_params}{jsdmStanFit}(object, ...)

\method{log_posterior}{jsdmStanFit}(object, ...)

\method{rhat}{jsdmStanFit}(object, ...)

\method{neff_ratio}{jsdmStanFit}(object, ...)
}
\description{
This is the jsdmStanFit class, which stan_gllvm and stan_mglmm both create.
}
\section{Functions}{
\itemize{
\item \code{print.jsdmStanFit}: Print the default summary for the model

\item \code{get_parnames}: Get the model parameter names

\item \code{nuts_params.jsdmStanFit}: Get the NUTS parameters

\item \code{log_posterior.jsdmStanFit}: Get the log posterior

\item \code{rhat.jsdmStanFit}: Get the R-hat

\item \code{neff_ratio.jsdmStanFit}: Get the n_eff ratio
}}

\section{Elements for \code{jsdmStanFit} objects}{

\describe{
\item{\code{jsdm_type}}{
A length one character vector describing type of jSDM
}
\item{\code{family}}{
A character vector describing response family
}
\item{\code{species}}{
A character vector of the species names
}
\item{\code{sites}}{
character vector of the site IDs
}
\item{\code{preds}}{
A character vector of the measured predictors included
}
\item{\code{data_list}}{
A list containing the original data used to fit the model
(empty when save_data is set to FALSE)
}
\item{\code{n_latent}}{
A length one integer vector representing number of latent
variables (in gllvm type fits) or NA in all other cases
}
}
}

