% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo.R
\name{loo.jsdmStanFit}
\alias{loo.jsdmStanFit}
\alias{loo}
\title{Efficient approximate leave-one-out cross-validation using the \pkg{loo} package}
\usage{
\method{loo}{jsdmStanFit}(object, ...)
}
\arguments{
\item{object}{The jsdmStanFit model object}

\item{...}{Other arguments passed to the \code{\link[loo]{loo}} function}
}
\value{
A list with class \code{c("psis_loo","loo")}, as detailed in the
\code{\link[loo]{loo}} documentation
}
\description{
This function uses the \pkg{loo} package to compute PSIS-LOO CV, efficient
approximate leave-one-out (LOO) cross-validation for Bayesian models using Pareto
smoothed importance sampling (PSIS).
}
