% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{mcmc_plot.jsdmStanFit}
\alias{mcmc_plot.jsdmStanFit}
\alias{mcmc_plot}
\title{MCMC plots implemented in \pkg{bayesplot}}
\usage{
\method{mcmc_plot}{jsdmStanFit}(
  x,
  pars = NULL,
  type = "intervals",
  sample_n = 10,
  inc_warmup = FALSE,
  include = TRUE,
  regexp = FALSE,
  ...
)

mcmc_plot(x, ...)
}
\arguments{
\item{x}{The jsdmStanFit model object}

\item{pars}{The parameters to plot, by default a random sample of twenty of the
parameters fit within the model}

\item{type}{The MCMC plot type to be used}

\item{sample_n}{If \code{pars = NULL} then the number of random non-sigma
parameters to include (details in description)}

\item{inc_warmup}{Whether to include the warmup period in the plots, by default
\code{FALSE}}

\item{include}{Whether to include or exclude the parameters specified by pars, by
default \code{TRUE} (i.e. include)}

\item{regexp}{If pars should be treated as a regular expression for matching to
parnames, by default \code{FALSE}}

\item{...}{Other arguments to be passed to the MCMC plotting function}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} object that can be customised using
the \pkg{ggplot2} package
}
\description{
An interface for calling the MCMC plotting functions implemented in the
\pkg{bayesplot} package
}
