% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_check.R
\name{multi_pp_check}
\alias{multi_pp_check}
\title{Multiple pp_check plots per species}
\usage{
multi_pp_check(
  object,
  plotfun = "dens_overlay",
  species = NULL,
  ndraws = NULL,
  grid_args = list(...),
  ...
)
}
\arguments{
\item{object}{The jsdmStanFit model object}

\item{plotfun}{The ppc plot function to use, given as a character string. The
default is to call \link[bayesplot:PPC-distributions]{ppc_dens_overlay}. Can be
specified as either the entire name of function as a character string or without
the ppc_ prefix.}

\item{species}{Which species should be included, by default all}

\item{ndraws}{How many draws should be used within the plots}

\item{grid_args}{Optional list of arguments that are passed to
\code{\link[gridExtra:arrangeGrob]{gridExtra::arrangeGrob()}} (nrow, ncol, widths, etc.)}

\item{...}{Other options passed to pp_check}
}
\value{
An object of class \code{"bayesplot_grid"}, for more information see \code{\link[bayesplot:bayesplot_grid]{bayesplot::bayesplot_grid()}}
}
\description{
Multiple pp_check plots per species
}
