% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ordiplot}
\alias{ordiplot}
\title{Plotting an ordination plot for a GLLVM model}
\usage{
ordiplot(
  object,
  choices = c(1, 2),
  type = "species",
  ndraws = 30,
  draw_ids = NULL
)
}
\arguments{
\item{object}{The jsdmStanFit model object}

\item{choices}{Which latent variables to plot as dimensions, by default 1 and 2}

\item{type}{Whether to plot sites or species}

\item{ndraws}{How many draws to include in plot}

\item{draw_ids}{Which draws to include in plot (overrides ndraws)}
}
\value{
A ggplot2 object that can be modified using ggplot arguments
}
\description{
This function takes a GLLVM model fit and plots an ordination plot with a random
(or specified) selection of draws
}
\examples{

\donttest{
 # First simulate data and fit the model:
 gllvm_data <- jsdm_sim_data(method = "gllvm", N = 100, S = 6, D = 3,
                             family = "bernoulli")
 gllvm_fit <- stan_jsdm(dat_list = gllvm_data, method = "gllvm",
                        family = "bernoulli")

ordiplot(gllvm_fit)
# now plot the 1st and 3rd latent variables against each other for the sites:
ordiplot(gllvm_fit, choices = c(1,3), type = "sites")

}
}
