% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pinewood-data.R
\docType{data}
\name{pinewood}
\alias{pinewood}
\title{Vegetation data from native pinewoods in Scotland, 1971}
\format{
A data frame with 25 rows and 34 columns. Each column is a
species, or species aggregates and each row is a site. The site identity is marked
in the rownames of the data frame. Every entry is given as the number of
plots that species is observed in within the site (each site has 16 plots, other
than a small number of exceptions, detailed in pinewood_env.
}
\source{
\url{https://doi.org/10.5285/56a48373-771c-4d4a-8b5a-45ef496c6e55}
}
\usage{
pinewood
}
\description{
This is the plant data from a series of plots across native Scots Pine woodlands
in Scotland surveyed in 1971. It contains 34 plant species measured over 25 sites.
For further information on this data see Wood and Bunce (2016).
}
\details{
This data is made available under the terms of the Open Government Licence.
}
\references{
Wood, C. M., & Bunce, R. G. H. (2016). Ecological survey of the native
pinewoods of Scotland 1971. Earth System Science Data, 8(1), 177-189.
https://doi.org/10.5194/essd-8-177-2016
}
\seealso{
\code{\link{pinewood_env}}
}
\keyword{datasets}
