% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predict.R
\name{posterior_predict.jsdmStanFit}
\alias{posterior_predict.jsdmStanFit}
\alias{posterior_predict}
\title{Draw from the posterior predictive distribution}
\usage{
\method{posterior_predict}{jsdmStanFit}(
  object,
  newdata = NULL,
  newdata_type = "X",
  draws = NULL,
  list_index = "draws"
)
}
\arguments{
\item{object}{The model object}

\item{newdata}{New data, by default NULL and uses original data}

\item{newdata_type}{What form is the new data in, at the moment only
supplying covariates is supported.}

\item{draws}{Number of draws, by default the number of samples in the
posterior}

\item{list_index}{Whether to return the output list indexed by the number of
draws (default), species, or site.}
}
\value{
A list of linear predictors. If list_index is "draws" (the default)
the list will have length equal to the number of draws with each element of
the list being a site x species matrix. If the list_index is "species" the
list will have length equal to the number of species with each element of
the list being a draws x sites matrix. If the list_index is "sites" the
list will have length equal to the number of sites with each element of
the list being a draws x species matrix.
}
\description{
Draw from the posterior predictive distribution of the outcome.
}
