% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_check.R
\name{pp_check.jsdmStanFit}
\alias{pp_check.jsdmStanFit}
\title{Posterior predictive checks for \code{jsdmStanFit} objects}
\usage{
\method{pp_check}{jsdmStanFit}(
  object,
  plotfun = "dens_overlay",
  species = NULL,
  sites = NULL,
  summary_stat = "sum",
  calc_over = "site",
  ndraws = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class \code{jsdmStanFit}}

\item{plotfun}{The ppc plot function to use, given as a character string. The
default is to call \code{\link[bayesplot:PPC-distributions]{ppc_dens_overlay}}. Can be specified as}

\item{species}{Which species to include in the summary statistic, by default all}

\item{sites}{Which sites to include in the summary statistic, by default all}

\item{summary_stat}{The summary statistic to be used, by default \code{sum} but any
function can be used.}

\item{calc_over}{Whether to calculate the summary statistic by site or species, by
default \code{species}}

\item{...}{Additional arguments passed to \link[jsdmstan]{jsdm_statsummary}.}
}
\value{

}
\description{
Posterior predictive checks for \code{jsdmStanFit} objects
}
