% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_jsdm.R
\name{stan_jsdm}
\alias{stan_jsdm}
\alias{stan_mglmm}
\alias{stan_gllvm}
\title{Fit jsdm models in Stan}
\usage{
stan_jsdm(
  Y = NULL,
  X = NULL,
  species_intercept = TRUE,
  method,
  dat_list = NULL,
  family,
  site_intercept = FALSE,
  D = NULL,
  phylo = FALSE,
  covar = "matern_05",
  delta = 1e-05,
  save_data = TRUE,
  ...
)

stan_mglmm(
  Y = NULL,
  X = NULL,
  species_intercept = TRUE,
  dat_list = NULL,
  family,
  site_intercept = FALSE,
  phylo = FALSE,
  covar = "matern_05",
  delta = 1e-05,
  save_data = TRUE,
  ...
)

stan_gllvm(
  Y = NULL,
  X = NULL,
  D = NULL,
  species_intercept = TRUE,
  dat_list = NULL,
  family,
  site_intercept = FALSE,
  phylo = FALSE,
  covar = "matern_05",
  delta = 1e-05,
  save_data = TRUE,
  ...
)
}
\arguments{
\item{Y}{Matrix of species by sites. Rows are assumed to be sites, columns
are assumed to be species}

\item{X}{The covariates matrix, with rows being site and columns being
covariates}

\item{species_intercept}{Whether the model should be fit with an intercept by
species, by default TRUE}

\item{method}{Whether to fit a GLLVM or MGLMM model, details in description}

\item{dat_list}{Alternatively, data can be given to the model as a list containing
Y, X, N, S, K, and site_intercept. See output of mglmm_sim_data for an example of
how this can be formatted.}

\item{family}{The response family for the model, required to be one of
"gaussian", "bernoulli", "poisson" or "neg_binomial"}

\item{site_intercept}{Whether the model should be fit with a site intercept, by
default FALSE}

\item{D}{The number of latent variables within a GLLVM model}

\item{phylo}{A distance matrix between species (not necessarily phylogenetic). The
default FALSE does not incorporate phylogenetic information.}

\item{covar}{The covariance function as a character string, options are Matérn
kernel with \eqn{\nu} 1/2 ("matern_05"), 3/2 ("matern_15"), 5/2 ("matern_25"), or
infinite ("matern_inf"). Matérn kernel with infinite nu is equivalent to the
squared exponential kernel ("sq_exponential"), and with \eqn{\nu} = 1/2 the
exponential kernel ("exponential").}

\item{delta}{The constant added to the diagonal of the covariance matrix in the
phylogenetic model to keep matrix semipositive definite, by default 1e-5.}

\item{save_data}{If the data used to fit the model should be saved in the
model object, by default TRUE.}

\item{...}{Arguments passed to rstan::sampling}
}
\value{

}
\description{
Fit jsdm models in Stan
}
\section{Functions}{
\itemize{
\item \code{stan_mglmm}: Alias for stan_jsdm with method = "mglmm"

\item \code{stan_gllvm}: Alias for stan_jsdm with method = "gllvm"
}}

