% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_jsdm.R
\name{stan_mglmm}
\alias{stan_mglmm}
\alias{stan_mglmm.default}
\alias{stan_mglmm.formula}
\title{Alias for \code{stan_jsdm} with \code{method = "mglmm"}}
\usage{
stan_mglmm(X, ...)

\method{stan_mglmm}{default}(
  X = NULL,
  Y = NULL,
  species_intercept = TRUE,
  dat_list = NULL,
  family,
  site_intercept = FALSE,
  phylo = FALSE,
  covar = "matern_05",
  delta = 1e-05,
  save_data = TRUE,
  iter = 4000,
  ...
)

\method{stan_mglmm}{formula}(formula, data = list(), ...)
}
\arguments{
\item{X}{The covariates matrix, with rows being site and columns being covariates.
Ignored in favour of data when formula approach is used to specify model.}

\item{...}{Arguments passed to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}}}

\item{Y}{Matrix of species by sites. Rows are assumed to be sites, columns are
assumed to be species}

\item{species_intercept}{Whether the model should be fit with an intercept by
species, by default \code{TRUE}}

\item{dat_list}{Alternatively, data can be given to the model as a list containing
Y, X, N, S, K, and site_intercept. See output of jsdm_sim_data for an example of
how this can be formatted.}

\item{family}{The response family for the model, required to be one of
\code{"gaussian"}, \code{"bernoulli"}, \code{"poisson"} or \code{"neg_binomial"}}

\item{site_intercept}{Whether the model should be fit with a site intercept, by
default \code{FALSE}}

\item{phylo}{A distance matrix between species (not necessarily phylogenetic). The
default \code{FALSE} does not incorporate phylogenetic information.}

\item{covar}{The covariance function as a character string, options are Matérn
kernel with \eqn{\nu} 1/2 (\code{"matern_05"}), 3/2 (\code{"matern_15"}), 5/2
(\code{"matern_25"}), or infinite (\code{"matern_inf"}). Matérn kernel with
infinite nu is equivalent to the squared exponential kernel
(\code{"sq_exponential"}), and with \eqn{\nu} = 1/2 the exponential kernel
(\code{"exponential"}).}

\item{delta}{The constant added to the diagonal of the covariance matrix in the
phylogenetic model to keep matrix semipositive definite, by default 1e-5.}

\item{save_data}{If the data used to fit the model should be saved in the model
object, by default TRUE.}

\item{iter}{A positive integer specifying the number of iterations for each chain,
default 4000.}

\item{formula}{The formula of covariates that the species means are modelled from}

\item{data}{Dataframe or list of covariates.}
}
\description{
Alias for \code{stan_jsdm} with \code{method = "mglmm"}
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default

\item \code{formula}: Formula interface
}}

