% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsdmstanfit-class.R
\name{summary.jsdmStanFit}
\alias{summary.jsdmStanFit}
\title{Summarise the model fit and data structure and give summaries for the parameters}
\usage{
\method{summary}{jsdmStanFit}(
  object,
  prob_quantiles = c(0.15, 0.85),
  digit_summary = 3,
  prob_pars_only = FALSE,
  na_filter = TRUE
)
}
\arguments{
\item{object}{The model object}

\item{prob_quantiles}{The quantiles to summarise the parameter estimates with, by
default the 15\% and 85\% quantiles}

\item{digit_summary}{The number of digits to round the results to}

\item{prob_pars_only}{Whether to limit output to parameters with Rhat > 1.01 or
effective sample size < 500, by default FALSE}

\item{na_filter}{Whether to remove parameters with NAs in Rhat - this includes the
parameters fixed to zero or one, such as the upper triangle of the cholesky
factor of the correlation matrix. By default TRUE}
}
\description{
Summarise the model fit and data structure and give summaries for the parameters
}
